fine tune TextEntry to show text inside the widget add shade orientation to Theme structure

+++ include/BlendInt/TextEntry.hpp
@@ -86,7 +86,7 @@ namespace BlendInt {
 
 		size_t m_length;
 
-		size_t m_cursor_position;
+		int m_cursor_position;
 
 		/**
 		 * @brief The position to print text
